move scimstatus to server
