+++ server/lib/src/valueset/datetime.rs
@@ -1,7 +1,8 @@
 use crate::prelude::*;
 use crate::schema::SchemaAttribute;
+use crate::valueset::ScimResolveStatus;
 use crate::valueset::{DbValueSetV2, ValueSet};
-use kanidm_proto::scim_v1::server::ScimResolveStatus;
+
 use smolset::SmolSet;
 use time::OffsetDateTime;
