+++ server/lib/src/valueset/hexstring.rs
@@ -1,7 +1,8 @@
 use crate::prelude::*;
 use crate::schema::SchemaAttribute;
+use crate::valueset::ScimResolveStatus;
 use crate::valueset::{DbValueSetV2, ValueSet};
-use kanidm_proto::scim_v1::server::ScimResolveStatus;
+
 use std::collections::BTreeSet;
 
 #[derive(Debug, Clone)]
