+++ server/lib/src/valueset/json.rs
@@ -1,8 +1,9 @@
 use crate::prelude::*;
 use crate::schema::SchemaAttribute;
+use crate::valueset::ScimResolveStatus;
 use crate::valueset::{DbValueSetV2, ValueSet};
 use kanidm_proto::internal::Filter as ProtoFilter;
-use kanidm_proto::scim_v1::server::ScimResolveStatus;
+
 use smolset::SmolSet;
 
 #[derive(Debug, Clone)]
