+++ server/lib/src/valueset/jws.rs
@@ -1,10 +1,10 @@
 use crate::prelude::*;
 use crate::schema::SchemaAttribute;
+use crate::valueset::ScimResolveStatus;
 use crate::valueset::{DbValueSetV2, ValueSet};
 use base64urlsafedata::Base64UrlSafeData;
 use compact_jwt::{crypto::JwsRs256Signer, JwsEs256Signer, JwsSigner};
 use hashbrown::HashSet;
-use kanidm_proto::scim_v1::server::ScimResolveStatus;
 
 #[derive(Debug, Clone)]
 pub struct ValueSetJwsKeyEs256 {
