+++ server/lib/src/valueset/nsuniqueid.rs
@@ -1,8 +1,9 @@
 use crate::prelude::*;
 use crate::schema::SchemaAttribute;
 use crate::value::NSUNIQUEID_RE;
+use crate::valueset::ScimResolveStatus;
 use crate::valueset::{DbValueSetV2, ValueSet};
-use kanidm_proto::scim_v1::server::ScimResolveStatus;
+
 use smolset::SmolSet;
 
 #[derive(Debug, Clone)]
