+++ server/lib/src/valueset/oauth.rs
@@ -1,3 +1,4 @@
+use crate::valueset::ScimResolveStatus;
 use std::collections::btree_map::Entry as BTreeEntry;
 use std::collections::{BTreeMap, BTreeSet};
 
@@ -8,8 +9,8 @@ use crate::utils::str_join;
 use crate::value::{OauthClaimMapJoin, OAUTHSCOPE_RE};
 use crate::valueset::{uuid_to_proto_string, DbValueSetV2, ValueSet};
 
+use kanidm_proto::scim_v1::server::ScimOAuth2ClaimMap;
 use kanidm_proto::scim_v1::server::ScimOAuth2ScopeMap;
-use kanidm_proto::scim_v1::server::{ScimOAuth2ClaimMap, ScimResolveStatus};
 
 #[derive(Debug, Clone)]
 pub struct ValueSetOauthScope {
