+++ server/lib/src/valueset/spn.rs
@@ -1,7 +1,7 @@
 use crate::prelude::*;
 use crate::schema::SchemaAttribute;
-use crate::valueset::{DbValueSetV2, ValueSet};
-use kanidm_proto::scim_v1::server::ScimResolveStatus;
+use crate::valueset::{DbValueSetV2, ScimResolveStatus, ValueSet};
+
 use smolset::SmolSet;
 
 #[derive(Debug, Clone)]
