+++ server/lib/src/valueset/utf8.rs
@@ -1,8 +1,8 @@
 use crate::prelude::*;
 use crate::schema::SchemaAttribute;
 use crate::utils::trigraph_iter;
-use crate::valueset::{DbValueSetV2, ValueSet};
-use kanidm_proto::scim_v1::server::ScimResolveStatus;
+use crate::valueset::{DbValueSetV2, ScimResolveStatus, ValueSet};
+
 use std::collections::BTreeSet;
 
 #[derive(Debug, Clone)]
